fix(api): align config and contract structure
