+++ .github/workflows/pr.yml
@@ -82,6 +82,7 @@ jobs:
         run: |
           set -euo pipefail
           mkdir -p target/llvm-cov
+          cargo +nightly llvm-cov clean --workspace
           cargo +nightly llvm-cov nextest \
             --workspace \
             --profile ci \
